Change to explicitly return 0 for SH_DialogButtonBox_ButtonsHaveIcons
Icons on dialog buttons (generally only happens on Linux) tend to give
poor and inconsistent results when both light and dark themes can be
chosen independent of what the QPA offers for icons. None of the other
platforms have icons in dialog buttons (Mac, Windows, BeOS/Haiku...),
and they don't seem to offer much in way of usability, especially since
much of the time a misleading or no icon may be used on some buttons.