Fix potential floating point divide-by-zero in hsluv-c
Reported by @jcelerier. This is only a noticeable problem when using
UBsan, since it will detect floating point division by zero. While this
is not an error under IEEE floating point rules (it should produce a
NaN), clang and gcc pretend that they do not officially support IEEE
floating point in the strict general case so that they can do some
optimizations. However, they will not exploit floating point division-
by-zero to do anything nefarious like delete code paths. (This would
break a lot of software.)

The extra test/branch here is probably only about 20% slower on Core 2
Duo and this code is not in a hot path, so it's worth fixing if nothing
else just to stop UBsan warnings for people who use it.