Fix linking error when built with Qt-in-a-namespace
`qt_defaultDpiX()` is not forward declared by Qt headers, so it's not
resolved unless we explicitly say that it's in the Qt user namespace,
if it exists.

PhantomStyle still relies on qglobal.h placing QT_USE_NAMESPACE into the
header files. If QT_NO_USING_NAMESPACE is defined then PhantomStyle
won't build. This could be fixed by explicitly doing a QT_USE_NAMESPACE
in the Phantom namespace, and putting the name-prepend wrappers around
the Qt namespaced types in the phantomstyle headers, but I don't have
an environment set up to test this right now to make sure it works, and
I don't believe that configuration is commonly used by projects.