Add /wd5045 to disable /Qspectre hints implied by /Wall in VS2017 15.5.5
This occurs in updated versions of VS2017 and later. The warnings are
hints as to what would occur if /Qspectre was enabled. Building
phantomstyleplugin and funhouse should not emit a bunch of irrelevant
/Qspectre warnings if the user has not specified /Qspectre, so we
disable the warning.

This only happens to phantomstyleplugin and funhouse because the .pro
file for them enables /Wall by default.