Add optimistic fast-path for palette reuse
This commit adds an optimistic fast-path for getting a PhSwatch derived
from a QPalette when the QPalette hasn't changed. Specifically, we add a
path where we check the combination of QPalette::cacheKey() plus
QPalette::currentColorGroup() against the last such combination we had,
and if it's a match, then we know we can use the head of the cache list
as the PhSwatch for that QPalette.

If it doesn't match, then we use the more accurate hashing and
comparison as normal, and look at the full set of cached PhSwatches as
normal, etc.

This saves a couple hundred nanoseconds on average, with only a minor
hit when it doesn't match. The common case is that it matches, so this
can save the work of doing the full accurate QPalette hash.

There is not much additional code bloat -- a possible extra function
call, a few bitwise operations, a branch, an extra function argument,
and an extra quint64 value on the heap.