Remove unnecessary assignment in PhSwatch::loadFromQPalette()
For most kinds of palettes, this would have merely been redundant.
However, if your palette had QPalette::Text set as a different color
than QPalette::WindowText, then it would have used the wrong color.
This bug was not visible in any of the test palettes or system palettes,
because QPalette::Text is usually the same color as
QPalette::WindowText. But that's not a guarantee.

This commit fixes the bug.