Fix useless extra-safety null ptr check in SH_Table_GridLineColor case
It's not always clear if style-calling code is required to pass the
`option` argument as non-null. Checks in the Qt code are inconsistent.
Qt's own code, for this SH_Table_GridLineColor case, never passes a null
pointer.

The original QFusionStyle code had this return statement:

return option ? option->palette.window().color().darker(120).rgb() : 0;

for the SH_Table_GridLineColor case in `QFusionStyle::styleHint()`.

The Phantom code unnecessarily preserved this `option ?` check.
Code earlier in the block already assumed that `option` was non-null.

The two options are to remove the unnecessary null check, or to make it
actually work. Even though Qt's own code never seems to pass a null
pointer here, it's better to assume that custom user code may do so, and
we should try to allow this to keep happening in existing projects
without breaking them.