Add workaround for QPushButton content and font metrics height mismatch
In high DPI (and probably some other situations), it's possible for the
font metrics height and the measured height of a line of text to be
mismatched slightly. We want QPushButton heights and QComboBox (and
also QLineEdit) heights to match, assuming they don't have any icons
with special sizes or multi-line text.

But in high DPI scaling, the height of a QPushButton might be too large
by 1 or 2 px (virtual), because QPushButton uses the measured size of
its contents text, and QComboBox uses the font metrics height.

Workaround: if the contents height of a QPushButton mismatches its font
metrics height by only 1 or 2 pixels, then use the font metrics height.
If the height differs by more than that, leave it alone.

This workaround could cause QPushButtons to have the wrong height by 1
or 2 (virtual) pixels if they use something like a specially-size icon
that just happens to make the contents height fall within that range,
even if the text is the correct height. But this workaround is probably
still worth it.